backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["colamini <1657619661@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.9"
uvicorn = "^0.34.2"
langchain = "^0.3.23"
faiss-cpu = "^1.10.0"
openai = "^1.75.0"
sentence-transformers = "^4.1.0"
python-dotenv = "^1.1.0"
langchain-community = "^0.3.21"
chromadb = "^1.0.5"
tiktoken = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
